+++ MicrosoftEntraID/azure_ad/user.py
@@ -12,7 +12,7 @@ from msgraph.generated.users.item.authentication.methods.item.reset_password.res
     ResetPasswordPostRequestBody,
 )
 from msgraph.generated.users.item.messages.messages_request_builder import MessagesRequestBuilder
-from pydantic.v1 import BaseModel
+from pydantic import BaseModel
 
 from .base import (
     MicrosoftGraphAction,
@@ -55,7 +55,7 @@ class GetUserAction(MicrosoftGraphAction):
     async def run(self, arguments: RequiredSingleUserArguments):
         query_params = MessagesRequestBuilder.MessagesRequestBuilderGetQueryParameters(
             # select requires a list of strings, so we need to convert the keys of the schema to a list
-            select=list(GetUserResults.schema()["properties"].keys()),
+            select=list(GetUserResults.model_json_schema()["properties"].keys()),
         )
         request_configuration = MessagesRequestBuilder.MessagesRequestBuilderGetRequestConfiguration(
             options=[ResponseHandlerOption(NativeResponseHandler())], query_parameters=query_params